Module Alpha_context.Delegate

This module re-exports definitions from Delegate_storage, Delegate_consensus_key, Delegate_missed_attestations_storage, Delegate_slashed_deposits_storage, Delegate_cycles, Delegate_rewards, and Forbidden_delegates_storage.

val check_not_tz4 : Tezos_protocol_environment_021_PsQuebec.Signature.public_key_hash -> unit Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult

val frozen_deposits_limit : context -> public_key_hash -> Tez.t option Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

val set_frozen_deposits_limit : context -> public_key_hash -> Tez.t option -> context Tezos_protocol_environment_021_PsQuebec.Lwt.t

val fold : context -> order:[ `Sorted | `Undefined ] -> init:'a -> f:(public_key_hash -> 'a -> 'a Tezos_protocol_environment_021_PsQuebec.Lwt.t) -> 'a Tezos_protocol_environment_021_PsQuebec.Lwt.t

val list : context -> public_key_hash list Tezos_protocol_environment_021_PsQuebec.Lwt.t

val drain : context -> delegate:public_key_hash -> destination:public_key_hash -> (context * bool * Tez.t * Receipt.balance_updates) Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

val cycle_end : context -> Cycle.t -> (context * Receipt.balance_updates * public_key_hash list) Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

val already_denounced : context -> public_key_hash -> Level.t -> Round.t -> Misbehaviour.kind -> bool Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

See Already_denounced_storage.already_denounced.

type reward_and_burn = {

reward : Tez.t;

amount_to_burn : Tez.t;

}

type punishing_amounts = {

staked : reward_and_burn;

unstaked : (Cycle.t * reward_and_burn) list;

}

val punish_double_signing : context -> operation_hash:Tezos_protocol_environment_021_PsQuebec.Operation_hash.t -> Misbehaviour.t -> public_key_hash -> Level.t -> rewarded:public_key_hash -> context Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

See Delegate_slashed_deposits_storage.punish_double_signing.

type level_participation =

| Participated

| Didn't_participate

val record_baking_activity_and_pay_rewards_and_fees : context -> payload_producer:public_key_hash -> block_producer:public_key_hash -> baking_reward:Tez.t -> reward_bonus:Tez.t option -> (context * Receipt.balance_updates) Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

val record_attesting_participation : context -> delegate:public_key_hash -> participation:level_participation -> attesting_power:int -> context Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

val current_frozen_deposits : context -> public_key_hash -> Tez.t Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

val initial_frozen_deposits : context -> public_key_hash -> Tez.t Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

val delegated_contracts : context -> public_key_hash -> Contract.t list Tezos_protocol_environment_021_PsQuebec.Lwt.t

See Contract_delegate_storage.delegated_contracts.

val registered : context -> public_key_hash -> bool Tezos_protocol_environment_021_PsQuebec.Lwt.t

val deactivated : context -> public_key_hash -> bool Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

val is_forbidden_delegate : t -> public_key_hash -> bool

See Forbidden_delegates_storage.is_forbidden.

val last_cycle_before_deactivation : context -> public_key_hash -> Cycle.t Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

See Delegate_activation_storage.last_cycle_before_deactivation.

module Consensus_key : sig ... end

val prepare_stake_distribution : context -> context Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

See Stake_storage.prepare_stake_distribution.

module Rewards : sig ... end

module Staking_parameters : sig ... end

module Shared_stake : sig ... end

module For_RPC : sig ... end

The functions in this module are considered too costly to be used in the protocol. They are meant to be used only to answer RPC calls.