Module Delegate.For_RPC

The functions in this module are considered too costly to be used in the protocol. They are meant to be used only to answer RPC calls.

type participation_info = {

expected_cycle_activity : int;

minimal_cycle_activity : int;

missed_slots : int;

missed_levels : int;

remaining_allowed_missed_slots : int;

expected_attesting_rewards : Tez.t;

}

val participation_info : context -> public_key_hash -> participation_info Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

val full_balance : context -> public_key_hash -> Tez.t Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

Returns the full 'balance' of the implicit contract associated to a given key, i.e. the sum of the spendable balance (given by balance or Contract_storage.get_balance) and of the frozen balance of the contract.

The frozen balance is composed of all frozen bonds associated to the contract (given by Contract_storage.get_frozen_bonds), all unstaked frozen deposits, and of the fraction of the frozen deposits that actually belongs to the delegate and not to its stakers.

Only use this function for RPCs: this is expensive.

val delegated_balance : context -> public_key_hash -> Tez.t Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

val staking_balance : context -> public_key_hash -> Tez.t Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

val min_delegated_in_current_cycle : context -> public_key_hash -> (Tez.t * Level_repr.t option) Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

val has_pending_denunciations : context -> public_key_hash -> bool Tezos_protocol_environment_021_PsQuebec.Lwt.t

val pending_denunciations : context -> public_key_hash -> Denunciations_repr.t Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

val pending_denunciations_list : context -> (public_key_hash * Denunciations_repr.item) list Tezos_protocol_environment_021_PsQuebec.Lwt.t

val get_estimated_shared_pending_slashed_amount : context -> public_key_hash -> Tez.t Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

get_estimated_shared_pending_slashed_amount ctxt delegate returns the estimated shared pending slashed amount of the given delegate according to the currently available denunciations.