Module Tezos_raw_protocol_021_PsQuebec.Block_payload_repr

Value on which validators try to reach a consensus.

Consensus at a given level is reached on a sequence of operations. However, to differentiate between two blocks having the same sequence of operations, assuming that could ever happen (for instance, two empty blocks), we also include the hash of the block that precedes the block where these operations should be included.

val hash : predecessor_hash:Tezos_protocol_environment_021_PsQuebec.Block_hash.t -> payload_round:Round_repr.t -> Tezos_protocol_environment_021_PsQuebec.Operation_list_hash.elt list -> Block_payload_hash.t

Create a payload hash from the predecessor block hash, the first round at which the payload was proposed, and the hashes of non-consensus operations.