Parameter Make_builder.C

include Tezos_protocol_environment_021_PsQuebec.Compare.COMPARABLE

type t

val compare : t -> t -> int

val compare_cost : t -> Saturation_repr.may_saturate Saturation_repr.t

compare_cost k returns the cost of comparing the given key k with another value of the same type.