Module Tezos_raw_protocol_021_PsQuebec.Commitment_repr

type t = {

blinded_public_key_hash : Blinded_public_key_hash.t;

amount : Tez_repr.t;

}

This type represents a commitment to an amount of tokens which can be claimed by a fund raiser after the blockchain is deployed.

val encoding : t Tezos_protocol_environment_021_PsQuebec.Data_encoding.t