Module Tezos_raw_protocol_021_PsQuebec.Dal_operations_repr

DAL operations

This module gathers the datatypes for the payloads of DAL operations.

This module ensures the consistency with the internal data-structures like Dal_slot_index_repr.t.

module Publish_commitment : sig ... end