Module Dal_slot_repr.Slot_market

Only one slot header is accepted per slot index. If two slots headers are included into a block, the second one will fail.

Consequently, we rely on the order of operations which is done thanks to the fee market.

This is encapsulated in the following module.

type t

Represent the fee market for a list of slots.

val init : length:int -> t

init ~length encodes a list of length slots without candidates.

val length : t -> int

length t returns the length provided at initialisation time (see init).

val register : t -> Header.t -> (t * bool) option

register t index fees updates the candidate associated to index index. Returns Some (_, true) if the candidate is registered. Returns Some (_, false) otherwise. Returns None if the index is not in the interval 0;length where length is the value provided to the init function.

val candidates : t -> Header.t list

candidates t returns a list of slot header candidates.