Module Tezos_raw_protocol_021_PsQuebec.Deposits_repr

type t = {

initial_amount : Tez_repr.t;

current_amount : Tez_repr.t;

}

Type representing unstaked frozen deposits.

initial_amount is the amount on which slashing should be based. current_amount is the current amount after slashing has happened.

There is a record per cycle.

The initial_amount may be increased during the current cycle only, when an unstake is requested.

val encoding : t Tezos_protocol_environment_021_PsQuebec.Data_encoding.t

val zero : t

val (+?) : t -> Tez_repr.t -> t Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult

val (++?) : t -> t -> t Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult