Module Tezos_raw_protocol_021_PsQuebec.Manager_counter_repr

Counters are used as anti-replay protection mechanism in manager operations: each manager account stores a counter and each manager operation declares a value for the counter. When a manager operation is applied, the value of the counter of its manager is checked and incremented.

include Tezos_protocol_environment_021_PsQuebec.Compare.S

type t

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool

val (>) : t -> t -> bool

val compare : t -> t -> int

val equal : t -> t -> bool

val max : t -> t -> t

val min : t -> t -> t

val init : t

Initial value for a counter (zero).

val succ : t -> t

Successor of a counter.

val pp : Tezos_protocol_environment_021_PsQuebec.Format.formatter -> t -> unit

Pretty-printer for counters.

val encoding_for_storage : t Tezos_protocol_environment_021_PsQuebec.Data_encoding.t

Encoding for a counter to be used in Storage.

val encoding_for_operation : t Tezos_protocol_environment_021_PsQuebec.Data_encoding.t

Encoding for a counter to be used in Operation_repr.

val encoding_for_RPCs : t Tezos_protocol_environment_021_PsQuebec.Data_encoding.t

Encoding for a counter to be used in RPCs.

val encoding_for_errors : t Tezos_protocol_environment_021_PsQuebec.Data_encoding.t

Encoding for a counter to be used in errors.

module Internal_for_injection : sig ... end

To be used in client injection only.

module Internal_for_tests : sig ... end