Module Tezos_raw_protocol_021_PsQuebec.Misbehaviour_repr

type kind =

| Double_baking

| Double_attesting

| Double_preattesting

Which double signing event has occurred.

type t = {

level : Raw_level_repr.t;

round : Round_repr.t;

kind : kind;

}

Internal representation of a double signing event used in Denunciations_repr.item.

For a double baking event, the level and round are those of both duplicate blocks. For a double (pre)attestating event, the level and round are those that appear in the Operation_repr.consensus_content of both duplicate consensus operations.

Note: the culprit pkh doesn't appear as a field here because it is typically used as a key when storing denunciation items in the context.

val kind_encoding : kind Tezos_protocol_environment_021_PsQuebec.Data_encoding.t

val encoding : t Tezos_protocol_environment_021_PsQuebec.Data_encoding.t

val compare_kind : kind -> kind -> int

Comparison function for double signing kinds.

Double_baking < Double_attesting < Double_preattesting

val equal_kind : kind -> kind -> bool

val compare : t -> t -> int

Comparison function for misbehaviours.

Misbehaviours are ordered by increasing level, then increasing round, then kind using compare_kind.