Module Tezos_raw_protocol_021_PsQuebec.Parameters_repr

This module defines protocol parameters, i.e. constants regulating the behaviour of the blockchain under the protocol.

type bootstrap_account = {

public_key_hash : Tezos_protocol_environment_021_PsQuebec.Signature.Public_key_hash.t;

public_key : Tezos_protocol_environment_021_PsQuebec.Signature.Public_key.t option;

amount : Tez_repr.t;

delegate_to : Tezos_protocol_environment_021_PsQuebec.Signature.Public_key_hash.t option;

consensus_key : Tezos_protocol_environment_021_PsQuebec.Signature.Public_key.t option;

}

An implict contract (account) initially existing on a chain since genesis.

type bootstrap_contract = {

delegate : Tezos_protocol_environment_021_PsQuebec.Signature.Public_key_hash.t option;

amount : Tez_repr.t;

script : Script_repr.t;

hash : Contract_hash.t option;
(*
If the contract hash is not provided, generate a fresh hash.
*)

}

An originated contract initially existing on a chain since genesis.

type bootstrap_smart_rollup = {

address : Sc_rollup_repr.Address.t;

pvm_kind : Sc_rollups.Kind.t;

boot_sector : string;

parameters_ty : Script_repr.lazy_expr;

whitelist : Sc_rollup_whitelist_repr.t option;

}

An originated smart rollup initially existing on a chain since genesis.

type t = {

bootstrap_accounts : bootstrap_account list;

bootstrap_contracts : bootstrap_contract list;

bootstrap_smart_rollups : bootstrap_smart_rollup list;

commitments : Commitment_repr.t list;

constants : Constants_parametric_repr.t;

security_deposit_ramp_up_cycles : int option;

no_reward_cycles : int option;

}

Protocol parameters define some constants regulating behaviour of the chain.

val bootstrap_account_encoding : bootstrap_account Tezos_protocol_environment_021_PsQuebec.Data_encoding.t

val encoding : t Tezos_protocol_environment_021_PsQuebec.Data_encoding.t

val check_params : t -> unit Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult