Module Sc_rollup_game_repr.Index

type t = private {

alice : Sc_rollup_repr.Staker.t;

bob : Sc_rollup_repr.Staker.t;

}

val to_path : t -> string list -> string list

to_path i p returns a new path with the path to the game indexed by i added as a prefix to path p. See Path_encoding module.

val of_path : string list -> t option

val path_length : int

val rpc_arg : t Tezos_protocol_environment_021_PsQuebec.RPC_arg.t

val encoding : t Tezos_protocol_environment_021_PsQuebec.Data_encoding.t

val compare : t -> t -> int

val make : Sc_rollup_repr.Staker.t -> Sc_rollup_repr.Staker.t -> t

val staker : t -> player -> Sc_rollup_repr.Staker.t

Given an index in normal form, resolve a given player (Alice or Bob) to the actual staker they represent.