Module Sc_rollup_inbox_merkelized_payload_hashes_repr.History

A History.t is a lookup table of merkelized_and_payloads. Payloads are indexed by their hash Hash.t. This history is needed in order to produce proof.

A subtlety of this history type is that it is customizable depending on how much of the inbox history you actually want to remember, using the capacity parameter. In the L1 we use this with capacity set to zero, which makes it immediately forget an old level as soon as we move to the next. By contrast, the rollup node uses a history that is sufficiently large to be able to take part in all potential refutation games occurring during the challenge period.

include Bounded_history_repr.S with type key = Hash.t and type value = merkelized_and_payload

type t

type key = Hash.t

module Map : Tezos_protocol_environment_021_PsQuebec.Map.S with type key = key

type value = merkelized_and_payload

type view = value Map.t

val empty : capacity:int64 -> t

empty ~capacity returns a new table whose maximum capacity is given.

val view : t -> view

Export a view of the given bounded cache

val encoding : t Tezos_protocol_environment_021_PsQuebec.Data_encoding.t

Encoding for values of type t

val pp : Tezos_protocol_environment_021_PsQuebec.Format.formatter -> t -> unit

Pretty-printer for values of type t

val find : key -> t -> value option

find key t returns Some value if there exists some value associated to key in the table, and None otherwise.

type Tezos_protocol_environment_021_PsQuebec.Error_monad.error +=

| Key_bound_to_different_value of {

key : key;

existing_value : value;

given_value : value;

}

val remember : key -> value -> t -> t Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult

remember key value t inserts a new entry (key |-> value) in t.

If key already exists in t, its associated binding value' should be equal to value. In this case, t is returned unchanged. Otherwise, an error Key_bound_to_different_value is returned.

If key is not already present in t, the new binding (key |-> value) is inserted in t. If the number of elements would exceed t's capacity after the insertion of the new binding, the oldest binding is removed from t.

The structure t is returned unchanged if its capacity is negative or null.

module Internal_for_tests : sig ... end

val no_history : t