Module PVM_with_context_and_state.Dal_with_history

val confirmed_slots_history : Dal_slot_repr.History.t

The reference/snapshot cell of the DAL skip list that stores confirmed slots.

val get_history : Dal_slot_repr.History.hash -> Dal_slot_repr.History.t option Tezos_protocol_environment_021_PsQuebec.Lwt.t

A function to retrieve a history from an underlying cache.

val page_info : (Dal_slot_repr.Page.content * Dal_slot_repr.Page.proof) option

In case we expect to generate an input proof that is a DAL page confirmation, we should provide via page_info the information of the page. That is: its content and the proof that the page is part of a confirmed slot whose ID is part of the page's ID.

In case we expect to generate an input proof to justify that a DAL page is not confirmed, the value of page_info should be None.

In case the proof doesn't involve DAL inputs, the value of page_info is None.

val dal_parameters : Dal_slot_repr.parameters

Some parameters of the DAL. Needed when checking a page's proof against a slot's Dal_slot_repr.commitment.

val dal_attestation_lag : int

The lag between the time an attestation is published on L1 (its published_level) and the level it should be confirmed.

val dal_number_of_slots : int

The number of DAL slots provided by the L1.

val dal_activation_level : Raw_level_repr.t option

The level at which the DAL got activated, None if the DAL has not yet been activated.

val dal_attested_slots_validity_lag : int