Module Sc_rollup_repr.Past_commitment_period

type t = {

commitment_period_in_blocks : int;
(*
The number of blocks separating two valid commitment for a given particular Tezos protocol.
*)

next_protocol_activation : Raw_level_repr.t;
(*
The activation level of the next protocol, which is also the last level during which the commitment period is valid. That is, if

inbox_level + commitment_period_in_blocks <= next_protocol_activation

then inbox_level + commitment_period_in_blocks is the expected level of the next commitment.

Otherwise, the commitment period of the next protocol applies.
*)

}

val encoding : t Tezos_protocol_environment_021_PsQuebec.Data_encoding.t