Module Tezos_raw_protocol_021_PsQuebec.Stake_context

Functions on stake and depending on the context.

val apply_limits : Raw_context.t -> Staking_parameters_repr.t -> Full_staking_balance_repr.t -> Stake_repr.t Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult

Apply the delegation_over_baking and staking_over_baking limits of a delegate. Overstaked tez count as delegated, overdelegated tez do not count at all.

val baking_weight : Raw_context.t -> Staking_parameters_repr.t -> Full_staking_balance_repr.t -> int64 Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult

The weight of a baker used for baking and attesting rights.

val optimal_frozen_wrt_delegated_without_ai : Raw_context.t -> Full_staking_balance_repr.t -> Tez_repr.t Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult