Module Tezos_raw_protocol_021_PsQuebec.Staking

type Tezos_protocol_environment_021_PsQuebec.Error_monad.error +=

| Cannot_stake_with_unfinalizable_unstake_requests_to_another_delegate

| Manual_staking_forbidden

val stake : Raw_context.t -> for_next_cycle_use_only_after_slashing:bool -> amount:[ `At_most of Tez_repr.t | `Exactly of Tez_repr.t ] -> sender:Tezos_protocol_environment_021_PsQuebec.Signature.Public_key_hash.t -> delegate:Tezos_protocol_environment_021_PsQuebec.Signature.public_key_hash -> (Raw_context.t * Receipt_repr.balance_updates) Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

stake ctxt ~sender ~for_next_cycle_use_only_after_slashing ~delegate amount add amount as sender's stake to delegate.

If for_next_cycle_use_only_after_slashing is true, the implicit finalisation is done for the next cycle. It is meant to be used only at cycle end after the application of the slashing.

val request_unstake : Raw_context.t -> for_next_cycle_use_only_after_slashing:bool -> sender_contract:Contract_repr.t -> delegate:Tezos_protocol_environment_021_PsQuebec.Signature.public_key_hash -> Tez_repr.t -> (Raw_context.t * Receipt_repr.balance_updates) Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

request_unstake ctxt ~for_next_cycle_use_only_after_slashing ~sender_contract ~delegate amount records a request from sender_contract to unstake amount from delegate.

If for_next_cycle_use_only_after_slashing is true, the unstake request and the implicit finalisation is done for the next cycle. It is meant to be used only at cycle end after the application of the slashing.

val finalize_unstake : Raw_context.t -> for_next_cycle_use_only_after_slashing:bool -> Contract_repr.t -> (Raw_context.t * Receipt_repr.balance_updates) Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

finalize_unstake ctxt ~for_next_cycle_use_only_after_slashing contract performs the finalization of all unstake requests from contract that can be finalized. An unstake request can be finalized if it is old enough, specifically the requested amount must not be at stake anymore and must not be slashable anymore, i.e. after consensus_rights_delay + max_slashing_period after the request. Amounts are transferred from the contract's delegate (at request time) unstaked frozen deposits to contract's spendable balance, minus slashing the requested stake undergone in between.

If for_next_cycle_use_only_after_slashing is true, the finalization is done for the next cycle. It is meant to be used only at cycle end after the application of the slashing.

type staking_automation =

| Auto_staking

| Manual_staking

Staking can be either automated or manual. If Adaptive Issuance is enabled, staking must be manual.

val staking_automation : Raw_context.t -> staking_automation

val check_manual_staking_allowed : Raw_context.t -> unit Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult