Module Tezos_raw_protocol_021_PsQuebec.Storage

Tezos Protocol Implementation - Typed storage

This module hides the hierarchical (key x value) database under pre-allocated typed accessors for all persistent entities of the tezos context.

This interface enforces no invariant on the contents of the database. Its goal is to centralize all accessors in order to have a complete view over the database contents and avoid key collisions.

module type Simple_single_data_storage = sig ... end

module Block_round : Simple_single_data_storage with type value = Round_repr.t

type missed_attestations_info = {

remaining_slots : int;

missed_levels : int;

}

module Slashed_deposits_history : sig ... end

module Unstake_request : sig ... end

module Contract : sig ... end

Storage from this submodule must only be accessed through the module `Contract`.

module Big_map : sig ... end

module Sapling : sig ... end

module Delegates : Storage_sigs.Data_set_storage with type t := Raw_context.t and type elt = Tezos_protocol_environment_021_PsQuebec.Signature.Public_key_hash.t

Set of all registered delegates.

module Consensus_keys : Storage_sigs.Data_set_storage with type t := Raw_context.t and type elt = Tezos_protocol_environment_021_PsQuebec.Signature.Public_key_hash.t

Set of all active consensus keys in cycle `current + consensus_rights_delay + 1`

module Pending_consensus_keys : Storage_sigs.Indexed_data_storage with type t := Raw_context.t * Cycle_repr.t and type key = Contract_repr.t and type value = Tezos_protocol_environment_021_PsQuebec.Signature.public_key

The pending consensus key of a delegate at the given cycle

module Pending_denunciations : Storage_sigs.Indexed_data_storage with type t := Raw_context.t and type key = Tezos_protocol_environment_021_PsQuebec.Signature.public_key_hash and type value = Denunciations_repr.t

All denunciations of the current and previous cycles that will have an effect (slashing, reward), i.e. all below 100%, deferred to the end of their slashing period.

module Slashed_deposits : Storage_sigs.Indexed_data_storage with type t := Raw_context.t and type key = Tezos_protocol_environment_021_PsQuebec.Signature.public_key_hash and type value = Slashed_deposits_history.t

History of slashed deposits: an associative list of cycles to slashed percentages.

type denounced = {

for_double_preattesting : bool;

for_double_attesting : bool;

for_double_baking : bool;

}

This type is used to track which denunciations have already been recorded, to avoid slashing multiple times the same event.

val default_denounced : denounced

denounced with all fields set to false.

module Already_denounced : Storage_sigs.Indexed_data_storage with type t := Raw_context.t * Cycle_repr.t and type key = (Raw_level_repr.t * Round_repr.t) * Tezos_protocol_environment_021_PsQuebec.Signature.Public_key_hash.t and type value = denounced

Set used to avoid slashing multiple times the same event

module Pending_staking_parameters : Storage_sigs.Indexed_data_storage with type t := Raw_context.t * Cycle_repr.t and type key = Contract_repr.t and type value = Staking_parameters_repr.t

module Stake : sig ... end

module Delegate_sampler_state : Storage_sigs.Indexed_data_storage with type key = Cycle_repr.t and type value = Raw_context.consensus_pk Sampler.t and type t := Raw_context.t

State of the sampler used to select delegates. Managed synchronously with Stake.Selected_distribution_for_cycle.

module Issuance_bonus : Storage_sigs.Indexed_data_storage with type key = Cycle_repr.t and type value = Issuance_bonus_repr.t and type t := Raw_context.t

Compounding reward bonus for Adaptive Issuance

module Issuance_coeff : Storage_sigs.Indexed_data_storage with type key = Cycle_repr.t and type value = Tezos_protocol_environment_021_PsQuebec.Q.t and type t := Raw_context.t

Multiplicative coefficient for rewards under Adaptive Issuance (Includes the bonus)

Votes

module Vote : sig ... end

module type FOR_CYCLE = sig ... end

Seed

module Seed_status : Simple_single_data_storage with type value = Seed_repr.seed_status

module Seed : sig ... end

Storage from this submodule must only be accessed through the module `Seed`.

Commitments

module Commitments : Storage_sigs.Indexed_data_storage with type key = Blinded_public_key_hash.t and type value = Tez_repr.t and type t := Raw_context.t

module Ramp_up : sig ... end

Ramp up rewards

module Pending_migration : sig ... end

module Liquidity_baking : sig ... end

module Adaptive_issuance : sig ... end

module Global_constants : sig ... end

A map of Script_repr.expr values, indexed by their hash (Script_expr_hash.t). Values from this map can be incorporated by any contract via the primitive Michelson_v1_primitives.H_constant.

module Ticket_balance : sig ... end

This module exposes a balance table for tracking ticket ownership. The table is a mapping from keys to values where the keys consist of a hashed representation of:

Tenderbake

module Tenderbake : sig ... end

module Sc_rollup : sig ... end

module Dal : sig ... end

module Zk_rollup : sig ... end

module Legacy : sig ... end