Module Contract.Total_supply

Stores the amount of tokens currently present on chain

type context = Raw_context.t

type value = Tez_repr.t

The type of the value

val mem : context -> bool Tezos_protocol_environment_021_PsQuebec.Lwt.t

Tells if the data is already defined

val get : context -> value Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

Retrieve the value from the storage bucket ; returns a Storage_error if the key is not set or if the deserialisation fails

val find : context -> value option Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

Retrieves the value from the storage bucket ; returns None if the data is not initialized, or Storage_helpers.Storage_error if the deserialisation fails

val init : context -> value -> Raw_context.t Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

Allocates the storage bucket and initializes it ; returns a Storage_errorExisting_key if the bucket exists

val update : context -> value -> Raw_context.t Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

Updates the content of the bucket ; returns a Storage_Error Missing_key if the value does not exist

val add : context -> value -> Raw_context.t Tezos_protocol_environment_021_PsQuebec.Lwt.t

Allocates the data and initializes it with a value ; just updates it if the bucket exists

val add_or_remove : context -> value option -> Raw_context.t Tezos_protocol_environment_021_PsQuebec.Lwt.t

When the value is Some v, allocates the data and initializes it with v ; just updates it if the bucket exists. When the value is None, deletes the storage bucket ; does nothing if the bucket does not exist.

val remove_existing : context -> Raw_context.t Tezos_protocol_environment_021_PsQuebec.Error_monad.tzresult Tezos_protocol_environment_021_PsQuebec.Lwt.t

Delete the storage bucket ; returns a Storage_error Missing_key if the bucket does not exist

val remove : context -> Raw_context.t Tezos_protocol_environment_021_PsQuebec.Lwt.t

Removes the storage bucket and its contents ; does nothing if the bucket does not exist