Module Storage.Tenderbake

module First_level_of_protocol : Storage_sigs.Single_data_storage with type t := Raw_context.t and type value = Raw_level_repr.t

First_level_of_protocol stores the level of the first block of this protocol.

module Attestation_branch : Storage_sigs.Single_data_storage with type value = Tezos_protocol_environment_021_PsQuebec.Block_hash.t * Block_payload_hash.t and type t := Raw_context.t

Attestation_branch stores a single value composed of the grandparent hash and the predecessor's payload (computed with the grandparent hash) used to verify the validity of attestations.

module Forbidden_delegates : Storage_sigs.Single_data_storage with type value = Tezos_protocol_environment_021_PsQuebec.Signature.Public_key_hash.Set.t and type t := Raw_context.t

Forbidden_delegates stores the set of delegates that are not allowed to bake or attest blocks.